+++ testcube_client/cli.py
@@ -45,8 +45,10 @@ parser.add_argument('-start', '--start-run',
 parser.add_argument('-finish', '--finish-run',
                     help='Finish a run, require xunit files.',
                     action='store_true')
-parser.add_argument('-reset', '--rerun-result',
-                    help='Rerun a result by id, require xunit files.')
+parser.add_argument('-task', '--handle-task',
+                    help='Handler pending task one by one.')
+parser.add_argument('-reset', '--reset-result',
+                    help='Reset a result by reset_id, require xunit files.')
 parser.add_argument('-x', '--xunit-files',
                     help='Specify the xunit xml results, e.g "**/result*.xml"')
 parser.add_argument('-i', '--result-files',
@@ -136,14 +138,18 @@ def main():
         action(business.finish_run,
                result_xml_pattern=args.xunit_files)
 
+    # when handle pending task
+    elif args.handle_task:
+        action(business.handle_task())
+
     # when reset a result
     elif args.rerun_result:
         if not args.xunit_files:
             logging.error('Must specify --xunit-files!')
             return -1
 
-        action(business.rerun_result,
-               result_id=args.rerun_result,
+        action(business.reset_result,
+               reset_id=args.rerun_result,
                result_xml_pattern=args.xunit_files)
 
     # when user need to upload result files: --result-files